+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Flash messages for controllers moved to i18n yml
+- Generator `avocado:views` copies views directory into application
 
 ## [0.6.0] - 2023-07-25